Remove tag field in AbstractTransaction and move to LegacyTransaction class.
